* hotfix/RAY-63301_Turnaround-field-displays-wrong-timeframe:
  RAY-63301 - Fixed the issue for the turnaround time to show the pointed values

+++ app/ray/events/models.py
@@ -70,7 +70,7 @@ class JobQuoteCreatedEvent(BaseModel):
     sl: Language
     tl: list[Language]
     service: str
-    turnaround_days: int
+    turnaround_days: float
     quote: QuoteInfo
